- Updated version to 2.0.6 - LANDMarkClassifier()._check_params() now returns type List[np.ndarray, np.ndarray] - Removed TransformerMixin from imports in LANDMark.py - Removed tensorflow dependencies - Removed unused imports to 'gc' and 'pandas' from lm_linear_clfs.py - Each type of classification model now has its own module - Random selection of 'alpha' for RidgeClassifier when samples are fewer than 6 - Neural network now uses PyTorch (AMP not yet enabled) - Test coverage improvement
