+++ pyproject.toml
@@ -4,7 +4,7 @@ requires = ["hatchling"]
 
 [project]
 name = "LANDMarkClassifier"
-version = "2.0.5"
+version = "2.0.6"
 authors = [
     {name = "Josip Rudar", email = "rudarj@uoguelph.ca"},
     {name = "Teresita M. Porter"},
@@ -39,8 +39,6 @@ dependencies = [
     "numpy == 1.23.5",
     "scikit-learn >= 1.1.2",
     "joblib >= 1.2.0",
-    "tensorflow_addons >= 0.17.1",
-    "tensorflow >= 2.9.1",
     "pandas >= 1.5.0",
     "scipy >= 1.8.1",
 ]
